perf(gang): share the vehicle-weapon builder and keep its result stable
getVehicleWeapons existed twice, byte-identical apart from the hardpoint
fallback label ('Loc. unknown' on the fighter card, 'Unknown' on the print
roster). The print copy was also defined and called inside the fighter map, so
it returned a fresh array on every render -- which defeated React.memo on
WeaponTable and, after the previous commit, its useMemo as well. Print re-renders
on each of its display toggles, so that is the surface where it cost the most.

Move it to utils/weapon-variants.ts with the fallback label as a parameter --
the two strings are preserved rather than silently unified -- and have the print
roster read from a Map built once per roster prop. The map is keyed on the
fighter object, not its id, because the print roster holds the same fighter once
per loadout. NO_WEAPONS gives the empty case a stable reference too.

Verified byte-identical across 460 rendered cases, now including the print
roster (both gang sizes, roster branch forced on identically in both runs) and
crew fighters with hardpoint-mounted vehicle weapons covering arcs, operated-by
and both fallback labels.

Note: the "(Master-crafted)" weapon-name branch inside the builder is currently
unreachable from these pages -- assembleGangFighters does not emit
is_master_crafted on vehicle equipment (app/lib/shared/gang-assembly.ts:541-549).
Moved verbatim rather than changed; flagged rather than silently dropped.
