+++ components/gang/fighter-card.tsx
@@ -2,6 +2,7 @@ import React, { useEffect, useLayoutEffect, useRef, useState, memo, useMemo, use
 import { useRouter } from 'next/navigation';
 import { StatsTable, StatsType } from '../ui/fighter-card-stats-table';
 import WeaponTable from './fighter-card-weapon-table';
+import { buildVehicleWeapons, NO_WEAPONS } from '@/utils/weapon-variants';
 import { Equipment } from '@/types/equipment';
 import { FighterProps, FighterEffect, Vehicle, VehicleEquipment, FighterSkills } from '@/types/fighter';
 import { hasSaveCharacteristic } from '@/types/edition';
@@ -251,52 +252,6 @@ const FighterCard = memo(function FighterCard({
     ? Math.ceil(compactNaturalHeight * compactScale) + COMPACT_CARD_VERTICAL_BORDER_PX
     : undefined;
 
-  // Move getVehicleWeapons function before its usage
-  const getVehicleWeapons = (vehicle: Vehicle | undefined) => {
-    if (!vehicle?.equipment) return [];
-
-    const hardpoints = (vehicle.effects?.['hardpoint'] || []) as FighterEffect[];
-    
-    return vehicle.equipment
-      .filter(item => item.equipment_type === 'weapon')
-      .map(weapon => {
-        const weaponFighterId = weapon.fighter_weapon_id || weapon.vehicle_weapon_id || weapon.equipment_id;
-        const matchedHardpoint = hardpoints.find(hp => hp.fighter_equipment_id === weaponFighterId);
-        const hpData = matchedHardpoint?.type_specific_data && typeof matchedHardpoint.type_specific_data !== 'string'
-          ? matchedHardpoint.type_specific_data
-          : undefined;
-
-        return {
-          fighter_weapon_id: weaponFighterId,
-          weapon_id: weapon.equipment_id,
-          weapon_name: weapon.is_master_crafted || weapon.master_crafted 
-            ? `${weapon.equipment_name} (Master-crafted)`
-            : weapon.equipment_name,
-          weapon_profiles: weapon.weapon_profiles?.map(profile => ({
-            ...profile,
-            range_short: profile.range_short,
-            range_long: profile.range_long,
-            strength: profile.strength,
-            ap: profile.ap,
-            damage: profile.damage,
-            ammo: profile.ammo,
-            acc_short: profile.acc_short,
-            acc_long: profile.acc_long,
-            traits: profile.traits || '',
-            id: profile.id,
-            profile_name: profile.profile_name,
-            is_master_crafted: (profile as any).is_master_crafted || !!weapon.master_crafted || !!weapon.is_master_crafted
-          })) || [],
-          cost: weapon.cost,
-          ...(hpData && {
-            hardpoint_location: (hpData.location && String(hpData.location).trim()) || 'Loc. unknown',
-            hardpoint_arcs: Array.isArray(hpData.arcs) ? hpData.arcs as string[] : undefined,
-            hardpoint_operated_by: (hpData.operated_by === 'crew' || hpData.operated_by === 'passenger') ? hpData.operated_by : undefined,
-          }),
-        };
-      }) as unknown as Weapon[];
-  };
-
   // Only calculate vehicle stats for crew members
   const vehicleStats = useMemo(() => {
     if (!isCrew) return null;
@@ -305,8 +260,8 @@ const FighterCard = memo(function FighterCard({
 
   // Get vehicle weapons only for crew members
   const vehicleWeapons = useMemo(() => {
-    if (!isCrew || !vehicle) return [];
-    return getVehicleWeapons(vehicle);
+    if (!isCrew || !vehicle) return NO_WEAPONS;
+    return buildVehicleWeapons(vehicle, 'Loc. unknown');
   }, [isCrew, vehicle]);
 
   const adjustedVehicleRules = useMemo(() => {
